+++ include/game/battle_impact_effect.h
@@ -102,7 +102,7 @@ BattleImpactParticleTask *BattleImpactParticle_CreateFromObject(
     BattleSceneObject *object, void (*callback)(BattleAITask *task));
 void BattleImpactTrailEmitter_Update(BattleAITask *task);
 BattleAITask *BattleImpactEmitter_Start(
-    u16 object_id, int mode, s16 target_scale_q4, u16 lifetime_multiplier,
+    u16 object_id, int mode, int target_scale_q4, int lifetime_multiplier,
     int spawn_interval, int size_multiplier, int target_size,
     int initial_z_offset, int z_velocity, int reserved);
